Raw connection made - initial processing
